[Refactor] 업무 공간이 없는 경우 컴포넌트로 분리
